+++ src/test/kotlin/data/repositories/AuthenticationRepositoryImplTest.kt
@@ -1,7 +1,8 @@
 package data.repositories
 
-import data.datasource.UserDataSource
-import data.mappers.toDto
+import data.csv_data.repositories.AuthenticationRepositoryImpl
+import data.csv_data.datasource.UserDataSource
+import data.csv_data.mappers.toDto
 import io.mockk.*
 import logic.models.User
 import logic.models.UserRole
